Implemented Flask app UI for chatbot
