--- nltk_utils.py
@@ -1,29 +1,0 @@
-from unicodedata import name
-import nltk
-nltk.download('punkt')
-import numpy as np
-from nltk.stem.porter import PorterStemmer
-
-stemmer = PorterStemmer()
-
-
-def tokenize(sentence):
-    return nltk.word_tokenize(sentence)
-
-def stem(word):
-    return stemmer.stem(word.lower())
-
-def bag_of_words(tokenized_sentence, all_words):
-    
-    tokenized_sentence = [stem(w) for w in tokenized_sentence]
-
-    # this will count how many times all words appear in the tokenized sentence, will print in the form a numpy array
-
-    bag = np.zeros(len(all_words), dtype=np.float32)
-    for idx, w in enumerate(all_words):
-        if w in tokenized_sentence:
-            bag[idx] = 1.0
-
-    return bag
-
-
