+++ templates/index.html
@@ -1,50 +1,59 @@
-<!DOCTYPE html>
-<html>
-<head>
-
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-
-
-
-</head>
-
-<body>
-
-<h1>Flask Chatbot</h1>
-
-<div>
-<div id="chatbox">
-<p class="botText"><span> Hi! I am your friendly Chatterbot.</span></p>
-</div>
-
-<div id="userInput">
-<input id ="textInput" type="text" name="msg" placeholder="Message">
-<input id ="buttonInput" type="submit" value="Send">
-</div>
-
-<script>
-function getBotResponse(){
-var rawText= $("#textInput").val();
-var userHtml = '<p class = "userText"><span>' + rawText + '</span></p>';
-$("#textInput").val("");
-$("#chatbox").append(userHtml);
-document.getElementById('userInput').scrollIntoView({block:'start',behaviour:'smooth'})
-$.get("/get", {msg:rawText}).done(function(data) {
-var botHtml='<p class="botText"><span>'+ data + '</span></p>';
-$("#chatbox").append(botHtml);
-document.getElementById('userInput').scrollIntoView({block:'start',behaviour:'smooth'})
-});
-}
-$("#textInput").keypress(function(e) {
-if(e.which == 13) {
-getBotResponse();
-}
-});
-$("#buttonInput").click(function() {
-getBotResponse();
-})
-</script>
-</div>
-
-</body>
-</html>
+<DOCTYPE html>
+    <html lang="en">
+        <head>
+            <meta charset="UTF-8">
+            <meta name="viewport" content="width=device-width, initial-scale=1.0">
+            <link rel="stylesheet" href="static/style.css" />
+        <!-- 	<link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/material-design-icons/3.0.1/iconfont/material-icon.min.css"> -->
+            <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/material-design-icons/3.0.1/iconfont/material-icons.min.css" integrity="sha512-mRuH7OxetZB1XiSaKZ2fxENKgxVvx3ffpzX0FUcaP6GBqAuqOLc8YiC/3EVTUVt5p5mIRT5D9i4LitZUQKWNCg==" crossorigin="anonymous" />
+            <title>Chat Window</title>
+    
+        </head>
+        <body>
+            
+    
+            <section>
+                <br>
+                <br>
+                <h1>Hi, I'm FoodieBot</h1>
+                <h3>Chatbot using PyTorch</h3>
+                <p>Click on the chat button to start chatting</p>
+                
+    
+                <button class="chat-btn">
+                    <!-- used google material icon for comment -->
+                    <i class="material-icons">
+                        comment
+                    </i>
+                </button>
+    
+                <!-- pop up -->
+                <div class="chat-popup">
+                    <!-- start red chat badge -->
+                    <div class="badge">1</div>
+    
+                    <!-- inside chat box area -->
+                    <div class="chat-area" id="chatId">
+                        <div class="income-msg">
+                            <!-- <img src="static/img/tom.jpeg" class="avatar" alt=""> -->
+                            <span class="msg">start typing....</span>
+                        </div>
+                    </div>
+    
+                    <!-- input area into chat -->
+                    <div class="input-area">
+                        <input type="text" tabindex="-1">
+                        <button class="submit">
+                            <i class="material-icons">
+                                send
+                            </i>
+                        </button>
+                    </div>
+                </div>
+            </section>
+    
+    
+            <!-- javascript libraries -->
+            <script type="text/javascript" src="static/main.js"></script>
+        </body>
+    </html>
